+++ dotnet/src/webdriver/Firefox/FirefoxDriverService.cs
@@ -219,18 +219,6 @@ namespace OpenQA.Selenium.Firefox
 
 
         /// <summary>
-        /// Creates a default instance of the FirefoxDriverService.
-        /// </summary>
-        /// <param name="options">Browser options used to find the correct GeckoDriver binary.</param>
-        /// <returns>A FirefoxDriverService that implements default settings.</returns>
-        [Obsolete("CreateDefaultService() now evaluates options in Driver constructor")]
-        public static FirefoxDriverService CreateDefaultService(FirefoxOptions options)
-        {
-            string fullServicePath = DriverFinder.FullPath(options);
-            return CreateDefaultService(Path.GetDirectoryName(fullServicePath), Path.GetFileName(fullServicePath));
-        }
-
-        /// <summary>
         /// Creates a default instance of the FirefoxDriverService using a specified path to the Firefox driver executable.
         /// </summary>
         /// <param name="driverPath">The directory containing the Firefox driver executable.</param>