+++ dotnet/test/common/DevTools/DevToolsSecurityTest.cs
@@ -6,7 +6,7 @@ using OpenQA.Selenium.Environment;
 
 namespace OpenQA.Selenium.DevTools
 {
-    using CurrentCdpVersion = V120;
+    using CurrentCdpVersion = V121;
 
     [TestFixture]
     public class DevToolsSecurityTest : DevToolsTestFixture