+++ java/src/org/openqa/selenium/grid/node/config/NodeOptions.java
@@ -117,7 +117,7 @@ public class NodeOptions {
     }
 
     Optional<String> hubAddress = config.get(NODE_SECTION, "hub");
-    if (!hubAddress.isPresent()) {
+    if (hubAddress.isEmpty()) {
       return Optional.empty();
     }
 
@@ -242,7 +242,7 @@ public class NodeOptions {
         ImmutableMultimap.builder();
 
     addDriverFactoriesFromConfig(sessionFactories);
-    addDriverConfigs(factoryFactory, sessionFactories);
+    addDriverConfigs(factoryFactory, sessionFactories, maxSessions);
     addSpecificDrivers(allDrivers, sessionFactories);
     addDetectedDrivers(allDrivers, sessionFactories);
 
@@ -358,7 +358,8 @@ public class NodeOptions {
 
   private void addDriverConfigs(
       Function<ImmutableCapabilities, Collection<SessionFactory>> factoryFactory,
-      ImmutableMultimap.Builder<Capabilities, SessionFactory> sessionFactories) {
+      ImmutableMultimap.Builder<Capabilities, SessionFactory> sessionFactories,
+      int maxSessions) {
 
     Multimap<WebDriverInfo, SessionFactory> driverConfigs = HashMultimap.create();
 
@@ -469,9 +470,7 @@ public class NodeOptions {
 
                     int driverMaxSessions =
                         Integer.parseInt(
-                            thisConfig.getOrDefault(
-                                "max-sessions",
-                                String.valueOf(info.getMaximumSimultaneousSessions())));
+                            thisConfig.getOrDefault("max-sessions", String.valueOf(maxSessions)));
                     Require.positive("Driver max sessions", driverMaxSessions);
 
                     WebDriverInfo driverInfoConfig =
@@ -532,7 +531,7 @@ public class NodeOptions {
   private void addSpecificDrivers(
       Map<WebDriverInfo, Collection<SessionFactory>> allDrivers,
       ImmutableMultimap.Builder<Capabilities, SessionFactory> sessionFactories) {
-    if (!config.getAll(NODE_SECTION, "driver-implementation").isPresent()) {
+    if (config.getAll(NODE_SECTION, "driver-implementation").isEmpty()) {
       return;
     }
 
@@ -563,7 +562,7 @@ public class NodeOptions {
             .filter(entry -> drivers.contains(entry.getKey().getDisplayName().toLowerCase()))
             .findFirst();
 
-    if (!first.isPresent()) {
+    if (first.isEmpty()) {
       throw new ConfigException("No drivers were found for %s", drivers.toString());
     }