+++ java/src/org/openqa/selenium/remote/codec/AbstractHttpCommandCodec.java
@@ -25,6 +25,7 @@ import static org.openqa.selenium.remote.DriverCommand.ADD_CREDENTIAL;
 import static org.openqa.selenium.remote.DriverCommand.ADD_VIRTUAL_AUTHENTICATOR;
 import static org.openqa.selenium.remote.DriverCommand.CANCEL_DIALOG;
 import static org.openqa.selenium.remote.DriverCommand.CLEAR_ELEMENT;
+import static org.openqa.selenium.remote.DriverCommand.CLICK_DIALOG;
 import static org.openqa.selenium.remote.DriverCommand.CLICK_ELEMENT;
 import static org.openqa.selenium.remote.DriverCommand.CLOSE;
 import static org.openqa.selenium.remote.DriverCommand.DELETE_ALL_COOKIES;
@@ -196,14 +197,14 @@ public abstract class AbstractHttpCommandCodec implements CommandCodec<HttpReque
     defineCommand(SET_USER_VERIFIED, post(webauthnId + "/uv"));
 
     // Federated Credential Management API
-    String fedcm = sessionId + "/fedcm";
-    defineCommand(CANCEL_DIALOG, post(fedcm + "/canceldialog"));
-    defineCommand(SELECT_ACCOUNT, post(fedcm + "/selectaccount"));
-    defineCommand(GET_ACCOUNTS, get(fedcm + "/accountlist"));
-    defineCommand(GET_FEDCM_TITLE, get(fedcm + "/gettitle"));
-    defineCommand(GET_FEDCM_DIALOG_TYPE, get(fedcm + "/getdialogtype"));
-    defineCommand(SET_DELAY_ENABLED, post(fedcm + "/setdelayenabled"));
-    defineCommand(RESET_COOLDOWN, post(fedcm + "/resetCooldown"));
+    defineCommand(CANCEL_DIALOG, post("/fedcm/canceldialog"));
+    defineCommand(SELECT_ACCOUNT, post("/fedcm/selectaccount"));
+    defineCommand(CLICK_DIALOG, post("/fedcm/clickdialogbutton"));
+    defineCommand(GET_ACCOUNTS, get("/fedcm/accountlist"));
+    defineCommand(GET_FEDCM_TITLE, get("/fedcm/gettitle"));
+    defineCommand(GET_FEDCM_DIALOG_TYPE, get("/fedcm/getdialogtype"));
+    defineCommand(SET_DELAY_ENABLED, post("/fedcm/setdelayenabled"));
+    defineCommand(RESET_COOLDOWN, post("/fedcm/resetcooldown"));
 
     defineCommand(GET_DOWNLOADABLE_FILES, get(sessionId + "/se/files"));
     defineCommand(DOWNLOAD_FILE, post(sessionId + "/se/files"));